Refactor academic department service and query builder

+++ src/app/modules/academicDepartment/academicDepartment.service.ts
@@ -1,3 +1,5 @@
+import httpStatus from 'http-status';
+import AppError from '../../errors/AppError';
 import { TAcademicDepartment } from './academicDepartment.interface';
 import { AcademicDepartment } from './academicDepartment.model';
 
@@ -14,6 +16,10 @@ const getAllAcademicDepartmentsFromDB = async () => {
 const getSingleAcademicDepartmentFromDB = async (id: string) => {
   const result =
     await AcademicDepartment.findById(id).populate('academicFaculty');
+
+  if (!result) {
+    throw new Error('Academic Department not found');
+  }
   return result;
 };
 
@@ -21,13 +27,24 @@ const updateAcademicDepartmentIntoDB = async (
   id: string,
   payload: Partial<TAcademicDepartment>,
 ) => {
+  // check if academic department exists
+  const academicDepartment = await AcademicDepartment.findById(id);
+  if (!academicDepartment) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Academic Department not found');
+  }
   const result = await AcademicDepartment.findByIdAndUpdate(id, payload, {
     new: true,
   });
+
   return result;
 };
 
 const deleteAcademicDepartmentFromDB = async (id: string) => {
+  // check if academic department exists
+  const academicDepartment = await AcademicDepartment.findById(id);
+  if (!academicDepartment) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Academic Department not found');
+  }
   const result = await AcademicDepartment.findByIdAndDelete(id);
   return result;
 };
